requirements.txt
acres==0.5.0
aiohappyeyeballs==2.4.0
aiohttp==3.10.6
aiosignal==1.3.1
altair==5.5.0
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
blinker==1.9.0
cachetools==6.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.2.1
colorama==0.4.6
configobj==5.0.9
configparser==7.2.0
dataclasses-json==0.6.7
distro==1.9.0
etelemetry==0.3.1
filelock==3.18.0
frozenlist==1.4.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
groq==0.11.0
grpcio==1.73.1
h11==0.14.0
h2==4.2.0
hpack==4.1.0
httpcore==1.0.5
httplib2==0.22.0
httpx==0.27.2
httpx-sse==0.4.1
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
langchain==0.3.26
langchain-community==0.3.27
langchain-core==0.3.68
langchain-groq==0.2.0
langchain-openai==0.3.27
langchain-qdrant==0.2.0
langchain-text-splitters==0.3.8
langsmith==0.4.4
looseversion==1.3.0
lxml==6.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
multidict==6.1.0
mypy_extensions==1.1.0
narwhals==1.46.0
networkx==3.5
nibabel==5.3.2
nipype==1.10.0
numpy==1.26.4
openai==1.93.3
orjson==3.10.7
packaging==24.1
pandas==2.3.1
pathlib==1.0.1
pillow==11.3.0
podman-compose==1.5.0
portalocker==2.10.1
protobuf==6.31.1
prov==2.1.1
puremagic==1.30
pyarrow==20.0.0
pydantic==2.9.2
pydantic-settings==2.10.1
pydantic_core==2.23.4
pydeck==0.9.1
pydot==4.0.1
PyMuPDF==1.26.3
pyparsing==3.2.3
pypdf==5.7.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==310
pyxnat==1.6.3
PyYAML==6.0.2
qdrant-client==1.14.3
rdflib==7.1.4
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rpds-py==0.26.0
scipy==1.16.0
simplejson==3.20.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.35
streamlit==1.46.1
tenacity==8.5.0
tiktoken==0.9.0
toml==0.10.2
tornado==6.5.1
tqdm==4.67.1
traits==7.0.2
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.2.3
watchdog==6.0.0
yarl==1.12.1
zstandard==0.23.0
